Add precisions about permissions and ownership
To avoid quota errors, specify permission and ownership options to add to files that are transfered to or created on Irene. In fact for any file and directory to be taken into account on the right quota it must both be owned by the group corresponding to the project (gen12462 in our case, instead of the unipdide which is for university Paris Diderot) and the permission must be set so that the file/directory inherits from the group priviliege ("s" option in chmod, corresponding to set the group id bit, or gidbit).